Renamed tvshow components to show components for simplicity and to avoid confusion later
