+++ src/app/app.module.ts
@@ -22,7 +22,7 @@ import { HomeComponent } from './home/home.component';
 import { FooterComponent } from './footer/footer.component';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { AuthInterceptorService } from './shared/auth/auth-interceptor.service';
-import { TvshowComponent } from './shared/tvshow/tvshow.component';
+import { ShowComponent } from './shared/show/show.component';
 import { StreamingServicesComponent } from './what-to-watch/streaming-services/streaming-services.component';
 import { WatchlistEditorComponent } from './watchlist/watchlist-editor/watchlist-editor.component';
 
@@ -38,7 +38,7 @@ import { WatchlistEditorComponent } from './watchlist/watchlist-editor/watchlist
     ShowResultsComponent,
     WhatToWatchComponent,
     WatchlistComponent,
-    TvshowComponent,
+    ShowComponent,
     NavbarComponent,
     StreamingServicesComponent,
     WatchlistEditorComponent
