+++ src/app/shared/shared.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { Router, RouterModule } from "@angular/router";
 
-import { TvshowComponent } from "./tvshow/tvshow.component";
+import { ShowComponent } from "./show/show.component";
 import { AuthComponent } from './auth/auth.component';
 import { NavbarComponent } from "./navbar/navbar.component";
