+++ src/app/shared/show/show.component.html
@@ -1,0 +1,9 @@
+<a href="#" class="list-group-item clearfix">
+  <div class="float-left">
+  <h4 class="list-group-item-heading">{{ show.title }}</h4>
+  <p class="list-group-item-text mb-0">{{ show.genre }}</p>
+</div>
+<div class="float-right">
+  <img src="{{ show.coverImg }}" alt="{{ show.title }}" class="img-responsive rounded" style="max-height:50px"/>
+</div>
+</a>
