+++ src/app/shared/show/show.model.ts
