--- src/app/shared/tvshow/tvshow.component.html
@@ -1,1 +1,0 @@
-<p>tvshow works!</p>
