--- src/app/shared/tvshow/tvshow.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-tvshow',
-  templateUrl: './tvshow.component.html',
-  styleUrls: ['./tvshow.component.css']
-})
-export class TvshowComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
