+++ src/app/watchlist/watchlist.component.html
@@ -1,7 +1,7 @@
-<div class="row justify-content-between">
+<div class="row mb-3">
   <div class="col-md-12">
-    <h1>My List</h1>
-    <app-show-list></app-show-list>
+    <h1>My Watchlist</h1>
+    <app-show *ngFor="let show of Show"></app-show>
   </div>
   <div class="col-md-12"></div>
 </div>
