+++ src/app/what-to-watch/show-results/show-results.component.html
@@ -12,6 +12,6 @@
         style="max-height: 50px"/>
       </div>
     </a>
-    <app-tvshow></app-tvshow>
+    <app-show></app-show>
   </div>
 </div>
