+++ src/app/what-to-watch/show-results/show-results.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Show } from 'src/app/shared/tvshow/tvshow.model';
+import { Show } from 'src/app/shared/show/show.model';
 
 @Component({
   selector: 'app-show-results',
